Fix array slice when using negative indexes
When calling `arr.slice(-1)` the output was getting `[undefined,
*rest]`. The desired behavior would be to get the last element. The
array should start slicing at the elemetn that many back from the end of
the array. In order to fix the issue if the start or end of the slice
are negative they are replaced by the length + (-num). This gives the
desired effect.

[#695]
Amos King @adkron <amos@binarynoggin.com>